Crate ruma_state_res

Re-exports

pub use event_auth::auth_check;

pub use event_auth::auth_types_for_event;

pub use room_version::RoomVersion;

Modules

event_auth

room_version

Enums

Error

Represents the various errors that arise when resolving state.

Traits

Event

Abstraction of a PDU so users can have their own PDU types.

Functions

lexicographical_topological_sort

Sorts the event graph based on number of outgoing/incoming edges.

resolve

Resolve sets of state events as they come in.

Type Definitions

Result

Result type for state resolution.

StateMap

A mapping of event type and state_key to some value T, usually an EventId.