Function ruma_state_res::event_auth::valid_membership_change

pub fn valid_membership_change<E: Event>(
    state_key: &str, 
    user_sender: &UserId, 
    content: Value, 
    prev_event: Option<Arc<E>>, 
    current_third_party_invite: Option<Arc<E>>, 
    auth_events: &StateMap<Arc<E>>
) -> Result<bool>

Does the user who sent this member event have required power levels to do so.

user - Information about the membership event and user making the request.

prev_event - The event that occurred immediately before the user event or None.

auth_events - The set of auth events that relate to a membership event. this is generated by calling auth_types_for_event with the membership event and the current State.